Fix segfault in cache revalidation
The filter program chain wasn't properly removed when a flow is removed,
resulting in a use after free error.

The code to remove the filter program chain was also in an incorrect
function; it should be deleted when the datapath flow rule is deleted.  It
was therefore deleted regardless of whether there are still references to
the datapath rule.

Signed-off-by: Paul Chaignon <paul.chaignon@orange.com>